# cv-job-matching/openapi.yaml
openapi: 3.1.0
info:
  title: Boussole API
  version: 1.0.0
  description: >
    API du MVP Boussole (assistant de candidature). Conventions transverses
    (auth cookie+CSRF, pagination curseur, idempotence, rate limiting) :
    voir 12-api-contracts.md. Erreurs au format RFC 9457.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
paths:
...
  /jobs:
    get:
      summary: Recherche d'offres (filtres SQL + full-text + rerank vectoriel)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: remote, in: query, schema: { type: array, items: { $ref: "#/components/schemas/RemotePolicy" } }, style: form, explode: true }
        - { name: contract, in: query, schema: { type: array, items: { $ref: "#/components/schemas/ContractType" } }, style: form, explode: true }
        - { name: salary_min, in: query, schema: { type: integer } }
        - { name: posted_since, in: query, schema: { type: integer, description: jours } }
        - { name: include_blocked, in: query, schema: { type: boolean, default: true } }
        - { name: include_hidden, in: query, schema: { type: boolean, default: false } }
        - { name: saved_only, in: query, schema: { type: boolean, default: false }, description: ne renvoyer que les offres sauvegardées }
        - { name: sort, in: query, schema: { type: string, enum: [match, date, relevance], default: match } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: Page de résultats
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/JobCard" }
                  next_cursor: { type: [string, "null"] }
  /jobs/{id}:
    get:
      summary: Détail d'une offre (sources et liens d'origine inclus)
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobDetail" }
        "404": { $ref: "#/components/responses/Problem" }
...
  /jobs/{id}/saved-state:
    put:
      summary: Sauvegarder ou masquer une offre
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state]
              properties:
                state: { type: string, enum: [saved, hidden] }
      responses: { "204": { description: OK } }
    delete:
      summary: Retirer la sauvegarde / le masquage
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "204": { description: OK } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: boussole_session }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    Limit: { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
    Cursor: { name: cursor, in: query, schema: { type: string } }
...
  responses:
    Problem:
      description: Erreur RFC 9457
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        trace_id: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
    SeniorityLevel: { type: [string, "null"], enum: [intern, junior, mid, senior, lead, principal, executive, null] }
    ContractType: { type: string, enum: [permanent, fixed_term, freelance, internship, apprenticeship, other] }
    RemotePolicy: { type: string, enum: [onsite, hybrid, full_remote] }
...
    JobCard:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        company_name: { type: string }
        locations: { type: array, items: { type: string } }
        remote: { type: [string, "null"] }
        contract: { type: [string, "null"] }
        salary_label: { type: [string, "null"], description: "'45–55 k€' ou null si non communiqué" }
        posted_at: { type: [string, "null"], format: date-time }
        match:
          type: [object, "null"]
          properties:
            score: { type: integer }
            confidence: { type: integer }
            low_data: { type: boolean }
            has_blocking: { type: boolean }
        saved_state: { type: [string, "null"], enum: [saved, hidden, null] }
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/JobCard"
        - type: object
          properties:
            description_text: { type: string }
            language: { type: string }
            status: { type: string, enum: [active, expired, withdrawn] }
            expires_at: { type: [string, "null"], format: date-time }
            seniority: { $ref: "#/components/schemas/SeniorityLevel" }
            skills_required: { type: array, items: { type: string } }
            skills_nice: { type: array, items: { type: string } }
            sources:
              type: array
              minItems: 1
              items:
                type: object
                required: [name, original_url]
                properties:
                  name: { type: string }
                  original_url: { type: string, format: uri }
                  posted_at: { type: [string, "null"], format: date-time }
...